+++ App/connect_four/lib/Screens/GameScreen/board.dart
@@ -3,7 +3,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
-import 'columns_builder.dart';
+import 'column.dart';
 
 class Board extends StatelessWidget {
   List<BoardColumn> _buildBoard() {
