+++ App/connect_four/lib/Screens/GameScreen/game_screen.dart
@@ -1,4 +1,4 @@
-import 'package:connect_four/Screens/GameScreen/board_builder.dart';
+import 'package:connect_four/Screens/GameScreen/board.dart';
 import 'package:flutter/material.dart';
 
 class GameScreen extends StatelessWidget {
